Add font rendering, readability, Vec2 usage, lines, vector math.
Implemented a basic way of rendering text, with a set font (Arial Rounded MT) that is loaded asynchronously.

Made some things a little more readable, like events and math.

Use Vec2's in most native Hummingbird functions.

Added line rendering, due to currently the only way of rendering being in a batch it is just a slim rectangle.

Added some more Math functions, both general and vector, there now also is a way to initialize static children of Math classes, and with that there is the addition of a static 'zero' and 'one' vector for all versions (Vec2, 3, 4), and also colors for the Vec4.