Add a 'fixedUpdate' function for physics.
The 'fixedUpdate' function, like you would find it in game engines like Unity, runs at a simulated 50 FPS.
This framerate is settable via the 'HB.fixedUpdateFrequency' property.
The function is most offently used for physics, as it runs consistently while the 'update' function runs at the refresh rate of the display.