Move batching code to renderer.
All of the batched methods have moved to the renderer class for simplicity.

Where you previously had to call 'HB.batch.drawColoredRectangle' you now to just call 'HB.colorRectangle'.

The max vertex and index counts are now also properties of the renderer and passed to the buffers via their respective Init methods.

And the testing  file was expanded to use all of the drawing methods available.