Fix texture switch bug.
When a render method with a texture filled the buffer, the batch would reset along with the textureIndex/Cache. The old textureIndex would be passed along however.

This has now been fixed by adding a 'flushBatchIfBufferFilled' method, it returns true if the batch was flushed, false otherwise.

This gives you a window to get the textureIndex again after the reset before submitting to the buffers.