Add option to return integers in 'HB.Math.seededRandom'.
The value was normalized to act more like the vanilla 'Math.random'.

You can now turn this off to get integers, or even get the full 32-bit integer by turning integers on and supplying no low or high.

This can be turned off for the whole class or per value.